updated routes adn added controller file
